Feat: 가이드북 상세페이지 메인화면 구현
