+++ src/app/map/guidebook/list/search/_domain/mocks/dummyPlaces.ts
